moves use departure data to separate file
